Tying interaction to the back and front end for Sophie

+++ src/visual/states/GuiUserOption.java
@@ -60,7 +60,7 @@ public class GuiUserOption {
 
 	private void setColor(Color frontEnd, ColorProperty color) {
 		//Needs to be fixed. Need to talk about chances to RGB color
-		RGBColor next = new RGBColor((int)frontEnd.getRed(), (int)frontEnd.getGreen(), (int)frontEnd.getBlue());
+		RGBColor next = new RGBColor(frontEnd.getRed(), frontEnd.getGreen(), frontEnd.getBlue());
 		color.set(next);
 	}
